+++ ui/mat/dropdown_list.go
@@ -7,7 +7,7 @@ import (
 	"github.com/mokiat/gomath/sprec"
 	"github.com/mokiat/lacking/ui"
 	co "github.com/mokiat/lacking/ui/component"
-	"github.com/mokiat/lacking/ui/mat/layout"
+	"github.com/mokiat/lacking/ui/layout"
 )
 
 var (
