+++ ui/mat/paper.go
@@ -4,7 +4,7 @@ import (
 	"github.com/mokiat/gomath/sprec"
 	"github.com/mokiat/lacking/ui"
 	co "github.com/mokiat/lacking/ui/component"
-	"github.com/mokiat/lacking/ui/mat/layout"
+	"github.com/mokiat/lacking/ui/layout"
 )
 
 var (
