+++ ui/mat/switch.go
@@ -2,7 +2,7 @@ package mat
 
 import (
 	co "github.com/mokiat/lacking/ui/component"
-	"github.com/mokiat/lacking/ui/mat/layout"
+	"github.com/mokiat/lacking/ui/layout"
 )
 
 // SwitchData holds the data for a Switch component.
